+++ unicode-data/lib/Unicode/Char/Case.hs
@@ -16,9 +16,7 @@
 module Unicode.Char.Case
     ( -- * Predicates
       isLowerCase
-    , isLower
     , isUpperCase
-    , isUpper
       -- * Case mappings
       -- $case
 
@@ -57,22 +55,13 @@ import qualified Unicode.Internal.Char.SpecialCasing.UpperCaseMapping as C
 -- It uses the character property
 -- <https://www.unicode.org/reports/tr44/#Lowercase Lowercase>.
 --
+-- See: 'Unicode.Char.Case.Compat.isLower' for the legacy predicate.
+--
 -- @since 0.3.0
 {-# INLINE isLowerCase #-}
 isLowerCase :: Char -> Bool
 isLowerCase = P.isLowercase
 
--- | Returns 'True' for lower-case characters.
---
--- It uses the character property
--- <https://www.unicode.org/reports/tr44/#Lowercase Lowercase>.
---
--- @since 0.1.0
-{-# INLINE isLower #-}
-{-# DEPRECATED isLower "Use isLowerCase instead. Note that the behavior of this function does not match base:Data.Char.isLower. See Unicode.Char.Case.Compat for behavior compatible with base:Data.Char." #-}
-isLower :: Char -> Bool
-isLower = P.isLowercase
-
 -- | Returns 'True' for upper-case characters.
 --
 -- It uses the character property
@@ -81,26 +70,13 @@ isLower = P.isLowercase
 -- Note: it does /not/ match title-cased letters. Those are matched using:
 -- @'Unicode.Char.General.generalCategory' c ==
 -- 'Unicode.Char.General.TitlecaseLetter'@.
+--
+-- See: 'Unicode.Char.Case.Compat.isUpper' for the legacy predicate.
 --
 -- @since 0.3.0
 {-# INLINE isUpperCase #-}
 isUpperCase :: Char -> Bool
 isUpperCase = P.isUppercase
-
--- | Returns 'True' for upper-case characters.
---
--- It uses the character property
--- <https://www.unicode.org/reports/tr44/#Uppercase Uppercase>.
---
--- Note: it does /not/ match title-cased letters. Those are matched using:
--- @'Unicode.Char.General.generalCategory' c ==
--- 'Unicode.Char.General.TitlecaseLetter'@.
---
--- @since 0.1.0
-{-# INLINE isUpper #-}
-{-# DEPRECATED isUpper "Use isUpperCase instead. Note that the behavior of this function does not match base:Data.Char.isUpper. See Unicode.Char.Case.Compat for behavior compatible with base:Data.Char." #-}
-isUpper :: Char -> Bool
-isUpper = P.isUppercase
 
 -- $case
 --